+++ sdk/resourcemanager/containerservicefleet/armcontainerservicefleet/autoupgradeprofiles_client.go
@@ -46,7 +46,7 @@ func NewAutoUpgradeProfilesClient(subscriptionID string, credential azcore.Token
 // BeginCreateOrUpdate - Create a AutoUpgradeProfile
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-05-02-preview
+// Generated from API version 2025-01-23-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - fleetName - The name of the Fleet resource.
 //   - autoUpgradeProfileName - The name of the AutoUpgradeProfile resource.
@@ -74,7 +74,7 @@ func (client *AutoUpgradeProfilesClient) BeginCreateOrUpdate(ctx context.Context
 // CreateOrUpdate - Create a AutoUpgradeProfile
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-05-02-preview
+// Generated from API version 2025-01-23-preview
 func (client *AutoUpgradeProfilesClient) createOrUpdate(ctx context.Context, resourceGroupName string, fleetName string, autoUpgradeProfileName string, resource AutoUpgradeProfile, options *AutoUpgradeProfilesClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AutoUpgradeProfilesClient.BeginCreateOrUpdate"
@@ -120,7 +120,7 @@ func (client *AutoUpgradeProfilesClient) createOrUpdateCreateRequest(ctx context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-05-02-preview")
+	reqQP.Set("api-version", "2025-01-23-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.IfMatch != nil {
@@ -138,7 +138,7 @@ func (client *AutoUpgradeProfilesClient) createOrUpdateCreateRequest(ctx context
 // BeginDelete - Delete a AutoUpgradeProfile
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-05-02-preview
+// Generated from API version 2025-01-23-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - fleetName - The name of the Fleet resource.
 //   - autoUpgradeProfileName - The name of the AutoUpgradeProfile resource.
@@ -165,7 +165,7 @@ func (client *AutoUpgradeProfilesClient) BeginDelete(ctx context.Context, resour
 // Delete - Delete a AutoUpgradeProfile
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-05-02-preview
+// Generated from API version 2025-01-23-preview
 func (client *AutoUpgradeProfilesClient) deleteOperation(ctx context.Context, resourceGroupName string, fleetName string, autoUpgradeProfileName string, options *AutoUpgradeProfilesClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AutoUpgradeProfilesClient.BeginDelete"
@@ -211,7 +211,7 @@ func (client *AutoUpgradeProfilesClient) deleteCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-05-02-preview")
+	reqQP.Set("api-version", "2025-01-23-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.IfMatch != nil {
@@ -223,7 +223,7 @@ func (client *AutoUpgradeProfilesClient) deleteCreateRequest(ctx context.Context
 // Get - Get a AutoUpgradeProfile
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-05-02-preview
+// Generated from API version 2025-01-23-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - fleetName - The name of the Fleet resource.
 //   - autoUpgradeProfileName - The name of the AutoUpgradeProfile resource.
@@ -274,7 +274,7 @@ func (client *AutoUpgradeProfilesClient) getCreateRequest(ctx context.Context, r
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-05-02-preview")
+	reqQP.Set("api-version", "2025-01-23-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -291,7 +291,7 @@ func (client *AutoUpgradeProfilesClient) getHandleResponse(resp *http.Response) 
 
 // NewListByFleetPager - List AutoUpgradeProfile resources by Fleet
 //
-// Generated from API version 2024-05-02-preview
+// Generated from API version 2025-01-23-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - fleetName - The name of the Fleet resource.
 //   - options - AutoUpgradeProfilesClientListByFleetOptions contains the optional parameters for the AutoUpgradeProfilesClient.NewListByFleetPager
@@ -339,7 +339,7 @@ func (client *AutoUpgradeProfilesClient) listByFleetCreateRequest(ctx context.Co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-05-02-preview")
+	reqQP.Set("api-version", "2025-01-23-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil